move waallet to packages
