+++ tabs/userOperationAuthorization.tsx
@@ -8,10 +8,10 @@ import { BackgroundDirectMessenger } from "~packages/messenger/background/direct
 import type { Paymaster } from "~packages/paymaster"
 import { NullPaymaster } from "~packages/paymaster/NullPaymaster"
 import { VerifyingPaymaster } from "~packages/paymaster/VerifyingPaymaster"
-import { WaalletContentProvider } from "~packages/provider/waallet/content/provider"
-import { WaalletRpcMethod } from "~packages/provider/waallet/rpc"
 import { ETH, Token } from "~packages/token"
 import json from "~packages/util/json"
+import { WaalletContentProvider } from "~packages/waallet/content/provider"
+import { WaalletRpcMethod } from "~packages/waallet/rpc"
 import type { Nullable } from "~typing"
 
 type PaymentOption = {
